Ubuntu: Remove fi, fl, ... ligatures
[why]
The fi and fl ligs are formulated in the font to go through glyphs with
codepoint `F001` and `F002`. That is detected and overwriting that
glyphs with some (Font Awesome) icons is prevented.

The question is: What is more important, to have the fi ligature or to
have the fa-music and fa-magnifying_glass icons?

[how]
We sacrify here Font Nerdism over Icon Nerdism and remove the
ligatures before patching.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>