Add SPDX license info
[why]
It can be hard to find out which license a particular font is using.
Therefore we add a new field to the fonts.json with an SPDX license
identifier (if possible).

[how]
For fonts with a license that has no SPDX identifier we use a free
`LicenseRef-*` identifier, but that reference is not solved in the
fonts.json file.

[note]
Also correct some licensing info given in the license audit.

Related: #1578

Suggested-by:  Jan Christian Gruenhage <jan.christian@gruenhage.xyz>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>