CI: Allow some tag to prevent locking old Issues
[why]
Suggested now fonts/glyphs/set are usually closed to keep the Issue list
to real Issues and not suggestions. People should be able to further
discuss that addition in the (closed) Issue and it will be evaluated on
the next release.

Anyhow, closing the Issue automatically locks it after half a year and
no discussion is possible anymore. That was not intended.

[how]
Add new tag that keeps closed Issues unlocked and mark the addition
Issues as such.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>