Update Firebase icon from original icon
[why]
Improve legibility. The full icon is in fact a multicolor icon and the
design depends on being rendered in multiple different colors. That is
impossible for fonts. So we should probably use their flat (mono-color)
icon instead, which hints the different 'flaps' through gaps instead of
colors.

[note]
Both icons (the old and the new) contradict rules given by Firebase to
use the logo. But that can not be helped in a icon font, as we just have
one color of undetermined hue.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>