Downloads: Make font name searchable
[why]
The download page is not searchable for font names (except in some cases
where the font is RFN or something).

Because the font name is just in the preview image (and thus an image)
users can not quickly go to some download link by searching for the
name.

[how]
Lay the font name invisible over the preview image.

Fixes: #1660

Reported-by: @willhansen
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>