Downloads: Increase invisible font name size
[why]
Some browsers (i.e. Safari) show the invisible text's highlighted when
searched for. The current size does not match the preview image's size
so the highlight box seems too small.

Interesting, Firefox and Chromium do not show the invisibe text box as
hit, which I personally always find very disruptive (when I search
something on a page and the brower finds it but nothing is
highlighted... where the heck did it find it?), but obviously Safari
does this right :-D

[how]
Just increase the (invisible) text size some bit to roughly fit.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>