Remove "curl" as from alternative download methods
[why]
The given link is wrong - and noone complained ;) So noone is using it.
Also we do not have all fonts anymore in the patched-font/ directory in
the repo.

I believe people will rather browse the repo in a browser and download
there. If they want to convert that into a curl call that is easily
possible. It would be very hard to use the given download method
directly.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>