doc: Show RFN permissions more visibly
[why]
The fonts where we are allowed to keep the font name despite the RFN
clause in the license have nothing mentioning of RFN or the exception in
the readmes. But it is recommended (required?) by SIL to inform users of
that exception in a most transparent way.

[how]
Add that bit of information to the readme files in the patched-fonts
directories and the release archives.

[note]
Also some other information seemed to be missing, by the complete
rebuild of the readme it also got updated (Gohu/ProFont).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>