FontAwesome: Scale glyphs a bit smaller
[why]
With Nerd Fonts v3.2.0 Font Awesome was updated to 6.5.1.

A lot icons got replaced and so the scaling was completely different
then with the old Font Awesome version as the icons were different.

Some complained that the new icons were to small, so that was fixed with
Nerd Fonts v3.2.1, where the size was increased by "approx 10%".

Now this seems a bit on the big side, so we decrease the size a bit.

[how]
v3.2.0   EM 1400    relative size 1.00
v3.2.1   EM 1200    relative size 1.17
v3.3.0   EM 1300    relative size 1.08

So we go to 92% of v3.2.1, i.e. -8% in height.

Note that the original change with v3.2.1 was +17% instead of the
intended 10% :-(

That 10% was just "some random number out of the blue".

Related: #1588
Fixes: #1614

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>