font-patcher: Fix weather icons scale for 'other' icons
[why]
A lot of icons that do not represent a 'weather state' but rather a
'weather event' are in the 'weather state' scaling group which scales
the icons waaaay down. This makes those icons almost unusable in stand
alone use cases.

[how]
Put all the extra icons into a new scaling group that is defined before
the weather state scaling group and thus takes precedence.

Split the final scaling group to have some icons on individual scale.

Fixes: #1708

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>