Devicons: Fix scaling of some few glyphs
[why]
Some glyphs in the (patched) E7B0-E7CF region seem to be too big.

[how]
The codepoints in the symbol font are 0x0100 less.
These are the "small things" that had a specific scale rule, which is
not needed anymore after we dropped these small icons and filled those
codepoints with new and bigger icons.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>