font-patcher: Fix 2 cell glyphs with overlap
[why]
Glyphs that shall end up in two cells with xy scaling and overlap (for
example E0C0, E0C2) end up being a bit too wide.
In fact the width is the 2 cell width and 2 overlap width.

[how]
Apply the overlap (which is 'percent' factor) not on the concrete glyph
width, which is 2*cellwidth in these cases, but base it on the
one-cell-width instead.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>