font-patcher: Fix alignment of some glyphs
[why]
The EE00 and EE03 glyphs are aligned too far to the left.

[how]
Affected are only right aligned glyphs with (negative) overlap, that
have an advance width (i.e. are member of a scale group).

Extract a new condition as variable as we need it in several places.
Use that condition also for the right alignment applicability check.

Related: #1733

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>