Trigger Release 3.3.0 (also: circumvent CI bug)
The CI first builds the fonts and THEN calls `generate-css`.
Which was ok, previously.

But now we store the glyphnames in the font files, and they are taken
from `glyphnames.json`, and THAT is only updated with the CSS file.

So the font files will get OLD names, if the symbol fonts got updated.

For now I just manually run the `glyphnames.json` update
(`generate-css.sh`) manually now, with this commit.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>