Update Sankey Diagram
[why]
With the latest release the numbers were just changed (in the svg) but
the flow graph stayed the same. The relative proportions can now look a
bit out of place.

[how]
Completely regenerate the graph from the Sankey generator.
Afterwards some manual adjustments to get the color and the position
right (in Inkscape).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>