doc: Mention needed bash version
[why]
Some bash scripts we have need bash version 4 or better. For example native
arrays where added with that version in 2009.

Unfortunately Apple decided to keep a very very old version of bash in
their setup which will not work.

[how]
Add information on needed bash version at least in the readme file.

We could now rewrite all scripts to work with bash 3; I tried that but that
is a tremendous amount of work and I guess the time is spent better on some
real problems; especially because only a small number of users are affected
and they could also just switch to recent bash version.

We could also add checks for the bash version to the scripts. This is not
done here.

[note]
Also remove obsolete explanatory entry; that subdir has been deleted.

Fixes: #1680

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>